requirements.txt
# Web Framework
streamlit>=1.28.0

# HTTP Requests & Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# PDF Processing
pypdf>=3.17.0

# Environment & Configuration
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9

# Machine Learning & AI
sentence-transformers>=2.0.0
scikit-learn>=1.0.0
numpy>=1.20.0
torch>=1.11.0
transformers>=4.41.0
huggingface-hub>=0.20.0

# Data Processing
pandas>=1.5.0

# Utilities
tqdm>=4.44.0
Pillow>=8.0.0
scipy>=1.6.0
typing-extensions>=4.5.0

